# admin-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "arctern",
    "description": "ZFS replication daemon HTTP API",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/jobs/{name}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cancel",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cancellation requested; the job remains active until the receiver releases the dataset"
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "Job kind does not support cancel"
          }
        }
      }
    },
    "/api/v1/jobs/{name}/pause": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "pause",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Transfer aborted resumably; scheduled cycles suspended"
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "Job kind does not support pause"
          }
        }
      }
    },
    "/api/v1/jobs/{name}/push/{peer}": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "push_to_peer",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "peer",
            "in": "path",
            "description": "Target peer name from the job's targets",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Manual replication to the peer queued"
          },
          "400": {
            "description": "Peer is not a target of this job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "Job kind does not support manual push"
          }
        }
      }
    },
    "/api/v1/jobs/{name}/resume": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "resume",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job unpaused; the next cycle resumes the partial transfer"
          },
          "404": {
            "description": "No such job"
          },
          "409": {
            "description": "Job kind does not support resume"
          }
        }
      }
    },
    "/api/v1/jobs/{name}/runs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "list_runs",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name as declared in arctern.toml",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Unix-second cutoff; rows with `started_at >= since` are returned.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rows to return. Defaults to 100, capped at 1000.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent job runs, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobRun"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{name}/wakeup": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "wakeup",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Job name as declared in arctern.toml",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job's cycle loop was woken up"
          },
          "404": {
            "description": "No job with that name is registered"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "Body shape for `4xx`/`5xx` responses from the daemon. `error` is a\nshort machine-readable category (`spawn`, `dataset_not_found`, …);\n`message` is a human-readable description.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "JobRun": {
        "type": "object",
        "description": "One row of `job_runs` returned by `GET /api/v1/jobs/{name}/runs`.\n`started_at` / `finished_at` are unix seconds; `bytes_sent` is set\nonly by push jobs that finished cleanly.",
        "required": [
          "started_at",
          "status"
        ],
        "properties": {
          "started_at": {
            "type": "integer",
            "format": "int64"
          },
          "finished_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "bytes_sent": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "RunStatus": {
        "type": "string",
        "description": "Terminal (or in-flight) state of one job cycle, as stored in\n`job_runs.status` and `push_syncs.status`.",
        "enum": [
          "ok",
          "error",
          "running",
          "cancelled",
          "dry_run",
          "interrupted"
        ]
      },
...
    }
  }
}